Dependencies: Switch to using the in-built Mach-O parser instead of otool
Now technically cross-platform except for the collector (which still
relies on install_name_tool) and also the fact that most paths
won't be resolvable, especially if it's not on the same filesystem.

This switch also helps speed up the calculation by ~ 30%-40% as we
remove the overhead of exec'ing another process and processing the
text from its output. otool by itself is still faster. duh. But it's
not recursive.